+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 0.2dev
 ------
 
+- Avoid division / 0 in ``bda.plone.discount.calculator.CartDiscount.vat``
+  [rnix]
+
 - Only hide border for global dicsount_views.
   [rnix]